move fed 1 directive definitions to ApolloGraphQL.HotChocolate.Federation.One namespace
