+++ src/Federation/Extensions/ApolloFederationSchemaBuilderExtensions.cs
@@ -1,4 +1,5 @@
 using ApolloGraphQL.HotChocolate.Federation;
+using ApolloGraphQL.HotChocolate.Federation.One;
 using AnyType = ApolloGraphQL.HotChocolate.Federation.AnyType;
 
 namespace HotChocolate;
@@ -32,6 +33,7 @@ public static class ApolloFederationSchemaBuilderExtensions
         builder.AddType<EntityType>();
         builder.AddType<ServiceType>();
         builder.AddType<FieldSetType>();
+        builder.AddType<ExtendsDirectiveType>();
         builder.AddType<ExternalDirectiveType>();
         builder.AddType<ProvidesDirectiveType>();
         builder.AddType<KeyDirectiveType>();
